Update test script

+++ test-cs.sh
@@ -10,7 +10,7 @@ num_iterations=$2
 
 registry_dir=`pwd`/default-ol/registry
 neutrino_registry_dir=`pwd`/../neutrino/registry
-log_file=$sandbox_type.log
+log_file=${sandbox_type}-$(date +%FT%T).log
 
 if [[ $sandbox_type == "graal" ]]; then
     file_ext=neutrino.so
@@ -30,19 +30,19 @@ if [[ ! -L $registry_dir ]]; then
     ln -s $neutrino_registry_dir $registry_dir
 fi
 
-rm -rf $log_file
-
 sum=0.0
 for i in `seq $num_iterations`; do
     func_name=${func}$i
     new_file=$func_name.$file_ext
-    cp $registry_dir/$ori_file $registry_dir/$new_file
+    if [[ ! -f $registry_dir/$new_file ]]; then
+        cp $registry_dir/$ori_file $registry_dir/$new_file
+    fi
 
     time_ms=$(python3 ./test-cs.py localhost 5000 $func_name $log_file)
 
     sum=$(echo "$sum+$time_ms" | bc)
 
-    rm -f $registry_dir/$new_file
+    #rm -f $registry_dir/$new_file
 done
 
 #echo "sum: $sum"
